Added files for wide case
Added new AUX case design for the larger MPU9250 and BNO08x boards, accommodations for both 25mm and 40mm straps as well as larger loops to more closely match the main case, allowing for more compatibility with different straps. Board is mounted sideways in case to allow for more natural wiring to the board. Has not been testing with actual boards but case prints wonderfully.